requirements.txt
flask>=3.0.0
neurokit2>=0.2.7
numpy>=1.24.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
matplotlib>=3.8.0
scipy>=1.11.0
Pillow>=10.0.0
